separated clock from speed slider
