+++ src/dashboard/src/Pages/Dashboard/components/ValueController.js
@@ -1,0 +1,77 @@
+import React, {useEffect, useState} from 'react';
+import useStyle from '../styles'
+import Slider from '@material-ui/core/Slider';
+import Input from '@material-ui/core/Input';
+
+//Value controller with slider + input.
+const ValueController = (props) => {
+    const classes = useStyle();
+    const [value, setValue] = useState(props.value);
+  
+    useEffect(() => {
+        setValue(props.value);
+    }, [props.value]); //runs when props.value updates.
+
+    useEffect(() => {
+        let a = props.onValueChange?.(value);
+    }, [value]); //runs when value updates.
+
+    const marks = [
+        {
+          value: 1,
+        }
+      ];
+
+    //onChange of slider
+    const handleSliderChange = (event, newValue) => {
+        setValue(newValue);
+    };
+  
+    //onChange of input
+    const handleInputChange = (event) => {
+        const targetValue = event.target.value === '' ? '' : Number(event.target.value);
+        setValue(targetValue);
+        let a = props.onValueChangeCommitted?.(event,targetValue);
+    };
+
+    //limit input
+    const handleBlur = () => {
+      if (value < 0) {
+        setValue(0);
+      } else if (value > 10) {
+        setValue(10);
+      }
+    };
+  
+    //render
+    return (
+      <div className={classes.root}>
+            <Slider
+              value={typeof value === 'number' ? value : 0}
+              onChange={handleSliderChange}
+              onChangeCommitted={(e,v)=>props.onValueChangeCommitted?.(e,v)}
+              aria-labelledby="input-slider"
+              step={0.1}
+              marks = {marks}
+              min={0}
+              max={10}
+              valueLabelDisplay="auto"
+            />
+            <Input
+              className={classes.input}
+              value={value}
+              margin="dense"
+              onChange={handleInputChange}
+              onBlur={handleBlur}
+              inputProps={{
+                step: 1,
+                min: 0,
+                max: 10,
+                type: 'number',
+                'aria-labelledby': 'input-slider',
+              }}
+            />
+      </div>
+    );
+}
+export default ValueController;
